requirements.txt
# MarketPulse - Market Sentiment & Price Movement Predictor
# Core
pandas>=2.0
numpy>=1.24
scipy>=1.10

# Data
yfinance>=0.2.31
pyyaml>=6.0

# ML
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
shap>=0.43

# Technical Analysis
pandas-ta>=0.3.14b1

# Visualization
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15

# Dashboard
streamlit>=1.28

# NLP / Sentiment (Phase 2 prep)
# transformers>=4.30
# torch>=2.0

# Testing
pytest>=7.4
pytest-cov>=4.1

# Utilities
tqdm>=4.65
joblib>=1.3
